Create Sum of second largest  numbers from even and odd locations of array
You are required to implement the following Function def large2sum(arr). 

The function accepts an integers arr of size ’length’ as its arguments you are required to return the sum of second largest element from the even positions and second largest from the odd position of given ‘arr’.

Coded in java.